bump nyxus dev version (#16)
* bump version up

* fix: added flag for slow tests

* fix: pre-commit hooks for remaining files

---------
