+++ features/nyxus-plugin/VERSION
@@ -1,1 +1,1 @@
-0.1.5-dev
+0.1.5-dev1
