+++ features/nyxus-plugin/src/polus/plugins/features/nyxus_plugin/__init__.py
@@ -1,2 +1,2 @@
 """Nyxus Plugin."""
-__version__ = "0.1.5-dev"
+__version__ = "0.1.5-dev1"
