+++ features/nyxus-plugin/tests/test_main.py
@@ -24,65 +24,6 @@ def clean_directories() -> None:
             shutil.rmtree(d)
 
 
-@pytest.fixture()
-def inp_dir() -> Union[str, Path]:
-    """Create directory for saving intensity images."""
-    return Path(tempfile.mkdtemp(dir=Path.cwd()))
-
-
-@pytest.fixture()
-def seg_dir() -> Union[str, Path]:
-    """Create directory for saving groundtruth labelled images."""
-    return Path(tempfile.mkdtemp(dir=Path.cwd()))
-
-
-@pytest.fixture()
-def output_directory() -> Union[str, Path]:
-    """Create output directory."""
-    return Path(tempfile.mkdtemp(dir=Path.cwd()))
-
-
-@pytest.fixture(params=[256, 512, 1024, 2048])
-def image_sizes(request: pytest.FixtureRequest) -> pytest.FixtureRequest:
-    """To get the parameter of the fixture."""
-    return request.param
-
-
-@pytest.fixture
-def synthetic_images(
-    inp_dir: Union[str, Path],
-    seg_dir: Union[str, Path],
-    image_sizes: pytest.FixtureRequest,
-) -> tuple[Union[str, Path], Union[str, Path]]:
-    """Generate random synthetic images."""
-    for i in range(10):
-        im = np.zeros((image_sizes, image_sizes))
-        points = image_sizes * np.random.random((2, 10**2))
-        im[(points[0]).astype(int), (points[1]).astype(int)] = 1
-        im = filters.gaussian(im, sigma=image_sizes / (20.0 * 10))
-        blobs = im > im.mean()
-        lab_blobs = measure.label(blobs, background=0)
-        intname = f"y04_r{i}_c1.ome.tif"
-        segname = f"y04_r{i}_c0.ome.tif"
-        int_name = Path(inp_dir, intname)
-        seg_name = Path(seg_dir, segname)
-        io.imsave(int_name, im)
-        io.imsave(seg_name, lab_blobs)
-    return inp_dir, seg_dir
-
-
-@pytest.fixture(
-    params=[
-        ("pandas", ".csv", "MEAN"),
-        ("arrowipc", ".arrow", "MEDIAN"),
-        ("parquet", ".parquet", "MODE"),
-    ]
-)
-def get_params(request: pytest.FixtureRequest) -> pytest.FixtureRequest:
-    """To get the parameter of the fixture."""
-    yield request.param
-
-
 def test_nyxus_func(
     synthetic_images: tuple[Union[str, Path], Union[str, Path]],
     output_directory: Union[str, Path],
@@ -116,12 +57,6 @@ def test_nyxus_func(
     clean_directories()
 
 
-@pytest.fixture(params=[5000, 10000, 30000])
-def scaled_sizes(request: pytest.FixtureRequest) -> pytest.FixtureRequest:
-    """To get the parameter of the fixture."""
-    yield request.param
-
-
 @pytest.fixture
 def scaled_images(
     inp_dir: Union[str, Path],
@@ -150,7 +85,7 @@ def get_scaled_params(request: pytest.FixtureRequest) -> pytest.FixtureRequest:
     yield request.param
 
 
-@pytest.mark.slow()
+@pytest.mark.skipif("not config.getoption('slow')")
 def test_scaled_nyxus_func(
     scaled_images: tuple[Union[str, Path], Union[str, Path]],
     output_directory: Union[str, Path],
